Addition of Readme file..
Great work Kanyisola.

I saw your readme file, but it cannot be easily viewed when someone visits this repo. I just re-added it in the normal format. kind merge this request and delete the previous readme file you created. This is still simply your documentation but in a different display mode.